fixes issue #16. nedit-ng now uses int64_t whereever appropriate!
I decided on int64_t instead of size_t because of a couple of reasons:

1. ensuring correctness when converting sign-sensitive code to unsigned is difficult to do correclty
2. if you manage to open a file that is > 2^63... I'm impressed, sorry we can't do that :-P

**However**, due to some built in limitations of Qt itself, files are limited to having 2^31 newlines in them.
This is because the scrollbars use an int type. I think that we have done a good job of lifting all practical
limmitations on filesize.

If a user manages to bump up against this limitation, we can revisit this issue