trimmed out the concept of having a complicatedseperate font system for italic, bold, and italic-bold styles. Instead, the more conventional methodology of having a single font set, which is styled as bold or italic as needed.
In other words, the user simply selects a single font/size for the
editor, and that font is style appropriately. This results in simpler
and more efficient rendering code as well!

For most (all?) users who just select a font and choose "fill from primary",
this will result in exactly the same as original behavior.

Closes issue #5.