fix errors on loading/unloading ctags file on Windows
The ':' (colon) character in the file path (for e.g, C:/path/to/file)
on Windows is mistreaded as path list, so nedit tries to load "C" and
"/path/to/file", where both would fail.

WARNING: this requires changing the path list separator of
nedit.tagFiles from ':' to ';' on Windows.