Well, had to change more than I expected for this one:
added a deque to manually track documents (instead of just iterating Qt's implicit lists of objects).

I did this because NEdit has concepts of "previous", "next", "first" and "last" documents, which are kinda
relative to app start time.

"first" is the first one create, "last" is the last one created.. it gets confusing because
"next" is simply the document that was created BEFORE you (it iterates down the singely linked list)
and "previous" (used in a single macro) mean to iterate the singley linked list backwards.

So...

To get the same behavior, switched to a deque, so we can push_front, this makes the iteration and indexing
not have to be logically reversed from the nedit5 version.

I'm gonna want to test this one a bunch to make sure I didn't mess anything up though :-/.