Mimicking NEdit-5's behavior... I think we should activeate the window too.
The original code did the following:

```
void RaiseWindow(Display *display, Window w, Boolean focus)
{
    if (focus) {
        XWindowAttributes winAttr;

        XGetWindowAttributes(display, w, &winAttr);
        if (winAttr.map_state == IsViewable)
            XSetInputFocus(display, w, RevertToParent, CurrentTime);
    }

    WmClientMsg(display, w, "_NET_ACTIVE_WINDOW", 0, 0, 0, 0, 0);
    XMapRaised(display, w);
}
```

So it appears to unconditionally raise and activcate the window, the only conditinal
is whether or not to set the INPUT focus to the "window" (document). So I think we are NOW
correct.

99% of the time, focus won't matter since the active tab (and therefore the document which
will get the focus by happenstance will be the one which would get the focus anyway.
So the "hidden" preference of `focusOnRaise` only has a tiny effect on the outcome.