Fixes issue #112!
So the problem was that by checking for file changes in the focus event itself, we are popping up
the messagebox BEFORE the focus event is complete. And the update caused by the focus event happens
AFTER it is complete.

So the solution is to NOT check for changes inside the focus event, but instead to schedule
a check for the immediate future. (Done via a single shot timer with a 0 timeout). Inspired
by kernel "bottom half interrupt handlers".

This allows the windowing system to complete processes of existing events, including the repaint
and then will immediately check for changes.

Woot!