Schedule scrolling for AFTER the widget is displayed (#242)
* annoyingly, on windows the viewrect isn't resized until it is shown
so if we try to do this now, it will try to scroll to a spot outside of
then "lineStarts_" array. The simplest thing to do to make it work correctly
is to schedule the scrolling for after the event is fully processed

Also, just because I noticed them in a debugger, initializing some variables
which are not initialized on construction. Probably harmless, but always best
to be safe.

* Some formatting because I haven't configured my windows editor to meet my preferences yet

* correcting intialization order