Instead of deleting menus immediately, schedule them for deletion in the near future. This fixes (hopefully) the main discussion of issue #249.
I believe the essential problem is that in some environments, deleting a menu
while processing an event triggered by a menu is no good. So if you have a macro
which happens to open a file (and thus trigger a menu rebuild), and you then execute
that macro from a menu... boom (on those platforms).

After reproducing the issue locally, I no longer see the crash after this patch!
Woot!